Implementing custom resource manager so that custom behaviour could be implemented for different resources kinds
