+++ src/services/crmanagers/OpenShiftWatchEvents.js
@@ -1,0 +1,9 @@
+const OpenShiftWatchEvents = Object.freeze({
+  MODIFIED: 'MODIFIED',
+  ADDED: 'ADDED',
+  DELETED: 'DELETED',
+  OPENED: 'OPENED',
+  CLOSED: 'CLOSED'
+});
+
+export { OpenShiftWatchEvents }
