+++ src/services/crmanagers/index.js
@@ -1,0 +1,4 @@
+import { OpenshiftResourceManagerFactory } from './OpenshiftResourceManagerFactory';
+import { OpenShiftWatchEvents } from './OpenShiftWatchEvents';
+
+export { OpenshiftResourceManagerFactory, OpenShiftWatchEvents };
